613: Building an alpine image containing the dcr binary and utils
The original Dockerfile produces a distroless image just containing the dcr binary.
In our toolchain it is more useful to have an image based on alpine so that we can run the dcr binary and do post processing on the output.

Updating the pipeline to use the new image and to execute a shell script rather than use the entrypoint. Parsing the dcr logs to find the client_id that gets created.

https://github.com/SecureBankingAccessToolkit/SecureBankingAccessToolkit/issues/613