811: Set expected error messages
Also added a step that can be used to output the x-fapi-interaction-id
of a response - handy when debugging and wanting to view logs in gcp for
a failing test case

One test is commented out and will need fixing at some point. A test to
check that if the registration jwt is signed using an unapproved
algorithm then we see a failure.

Issue: https://github.com/secureapigateway/secureapigateway/issues/811